Fix #894: revert egsphant to egs_brachy encoding
Revert to the egs_brachy encoding for egsphant files, allowing for 63
media encoded with the set of alphanumeric characters (where 0 is
reserved for vacuum).

Briefly, #633 expanded the number of media in egsphant files to 95,
using all printable ascii characters. However, the number of media was
also independently expanded in egs_glib for the development of
egs_brachy, but using a different encoding consisting of only the 63
alphanumeric ascii characters. This led to inconsistent egsphant files
that are no longer interchangeable.

This issue was originally reported and discussed in
https://github.com/clrp-code/egs_brachy/issues/22.